create node rediretcs separately
